refactor: split store/emails/hooks/hooks.ts in two separate hooks
