+++ src/views/search/extra-window/conversations/search-conversation-extra-window-panel.tsx
@@ -11,7 +11,7 @@ import { map } from 'lodash';
 
 import { Spinner } from 'assets/spinner';
 import { API_REQUEST_STATUS } from 'constants/index';
-import { useCompleteConversationOrFetch } from 'store/emails/hooks/hooks';
+import { useCompleteConversationOrFetch } from 'store/emails/hooks/use-complete-conversation-or-fetch';
 import { ConversationMessagePreviewWrapper } from 'views/app/detail-panel/conversation-message-preview-wrapper';
 import { SearchPanelHeader } from 'views/search/extra-window/search-panel-header';
