+++ src/views/search/panel/conversation/search-conversation-panel.tsx
@@ -15,7 +15,7 @@ import type {
 	SearchDetailPanelRouteParams
 } from '../../../../types/routes';
 import { API_REQUEST_STATUS, SEARCH_ROUTE } from 'constants/index';
-import { useCompleteConversationOrFetch } from 'store/emails/hooks/hooks';
+import { useCompleteConversationOrFetch } from 'store/emails/hooks/use-complete-conversation-or-fetch';
 import { SearchPanelHeader } from 'views/search/extra-window/search-panel-header';
 import { SearchConversationMessagePanel } from 'views/search/panel/conversation/search-conversation-message-panel';
